REF: Break metrics into submodules
Make the metrics module shorter by breaking it into topical
submodules.
