+++ src/xdesign/metrics/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""Objects and methods for computing the quality of reconstructions.
+
+.. moduleauthor:: Daniel J Ching <carterbox@users.noreply.github.com>
+"""
+
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from xdesign.metrics.coverage import *
+from xdesign.metrics.fullref import *
+from xdesign.metrics.standards import *
